# packages/contracts/openapi/v1/openapi.api.json
{
  "components": {
    "schemas": {
...
      "ApprovalList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApprovalRequest"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ApprovalList",
        "type": "object"
      },
...
      "ApprovalRequest": {
        "additionalProperties": false,
        "description": "人工确认（HIL）请求。人工确认是正式状态转换：审批解决后运行才能离开 WAITING_APPROVAL。",
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "decision_type": {
            "$ref": "#/components/schemas/DecisionType"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "evidence_snapshot_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "审批所依据的证据快照引用；Evidence 体系落地前允许为空。",
            "title": "Evidence Snapshot Id"
          },
          "expires_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Timestamp"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "pattern": "^appr_[0-9a-f]{32}$",
            "title": "Id",
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "minItems": 1,
            "title": "Options",
            "type": "array"
          },
          "resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Resolution"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "pattern": "^run_[0-9a-f]{32}$",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/omm_contracts__v1__approval_request__Status"
          },
          "step_id": {
            "anyOf": [
              {
                "pattern": "^step_[0-9a-f]{32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Id"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "decision_type",
          "title",
          "options",
          "status",
          "created_at"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "Budget": {
        "additionalProperties": false,
        "properties": {
          "cost_limit_usd": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Limit Usd"
          },
          "max_model_calls": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Model Calls"
          },
          "max_wall_time_s": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Wall Time S"
          }
        },
        "title": "Budget",
        "type": "object"
      },
      "BudgetInput": {
        "description": "与 schemas/v1/task-run.schema.json 的 budget 对象一致。",
        "properties": {
          "cost_limit_usd": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Limit Usd"
          },
          "max_model_calls": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Model Calls"
          },
          "max_wall_time_s": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Wall Time S"
          }
        },
        "title": "BudgetInput",
        "type": "object"
      },
...
      "CreateTaskRunInput": {
        "properties": {
          "auto_start": {
            "default": true,
            "title": "Auto Start",
            "type": "boolean"
          },
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "goal": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "workflow_version": {
            "default": "sim-0.1",
            "maxLength": 100,
            "minLength": 1,
            "title": "Workflow Version",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "goal"
        ],
        "title": "CreateTaskRunInput",
        "type": "object"
      },
...
      "Failure": {
        "additionalProperties": false,
        "properties": {
          "failure_class": {
            "$ref": "#/components/schemas/omm_contracts__v1__task_run__FailureClass"
          },
          "message": {
            "maxLength": 4000,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "failure_class",
          "message"
        ],
        "title": "Failure",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelingWorkspaceView": {
        "additionalProperties": false,
        "description": "建模运行面向 Web 的语义投影。后端只输出阶段、Agent 文案、动作和产物语义，前端负责映射到既有流程页面与 DOM 样式。",
        "properties": {
          "active_node": {
            "maxLength": 100,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Active Node",
            "type": "string"
          },
          "active_page": {
            "$ref": "#/components/schemas/PageKey"
          },
          "agent": {
            "$ref": "#/components/schemas/AgentProjection"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactProjection"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "goal": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "latest_event_sequence": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Sequence"
          },
          "pages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PageProjection"
                },
                "type": "array"
              },
              {
                "$ref": "#/components/schemas/Pages"
              }
            ],
            "maxLength": 6,
            "minLength": 6,
            "title": "Pages"
          },
          "pending_approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalProjection"
              },
              {
                "type": "null"
              }
            ]
          },
          "project_id": {
            "pattern": "^proj_[0-9a-f]{32}$",
            "title": "Project Id",
            "type": "string"
          },
          "project_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Project Name",
            "type": "string"
          },
          "run_id": {
            "pattern": "^run_[0-9a-f]{32}$",
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "suggested_route": {
            "$ref": "#/components/schemas/Route"
          },
          "updated_at": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "workflow_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Workflow Version",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "project_id",
          "project_name",
          "goal",
          "workflow_version",
          "run_status",
          "active_node",
          "active_page",
          "suggested_route",
          "agent",
          "pages",
          "artifacts",
          "pending_approval",
          "latest_event_sequence",
          "updated_at"
        ],
        "title": "ModelingWorkspaceView",
        "type": "object"
      },
...
      "NullableTimestamp": {
        "anyOf": [
          {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,6})?Z$",
            "type": "string"
          },
          {
            "type": "null"
          }
        ],
        "title": "NullableTimestamp"
      },
...
      "PageKey": {
        "enum": [
          "running",
          "data",
          "model",
          "experiments",
          "editor",
          "complete"
        ],
        "title": "PageKey",
        "type": "string"
      },
...
      "Pages": {
        "properties": {},
        "title": "Pages",
        "type": "object"
      },
...
      "RunId": {
        "pattern": "^run_[0-9a-f]{32}$",
        "title": "RunId",
        "type": "string"
      },
...
      "StepRunList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StepRun"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "StepRunList",
        "type": "object"
      },
...
      "TaskRun": {
        "additionalProperties": false,
        "description": "一次可暂停、恢复、重试、分支的 Agent 运行。status 是稳定生命周期枚举；current_node 是随 workflow_version 演进的领域阶段，消费方必须容忍未知节点名。",
        "properties": {
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Budget"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "current_node": {
            "maxLength": 100,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Current Node",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NullableTimestamp"
              },
              {
                "type": "null"
              }
            ]
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Failure"
              },
              {
                "type": "null"
              }
            ]
          },
          "goal": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/RunId"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "运行输入参数（题目引用、模拟钩子等），按 workflow_version 解释。",
            "title": "Params"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunId"
              },
              {
                "type": "null"
              }
            ],
            "description": "分支来源运行（规划文档 §7.1 的 parent_branch）。"
          },
          "project_id": {
            "pattern": "^proj_[0-9a-f]{32}$",
            "title": "Project Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NullableTimestamp"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/omm_contracts__v1__task_run__Status"
          },
          "updated_at": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "workflow_version": {
            "description": "工作流定义版本。首个模拟实现为 sim-0.1，节点集合：CREATED, PROBLEM_ANALYSIS, DATA_PREPARATION, MODEL_PLANNING, EXPERIMENTING, VALIDATING, PAPER_WRITING, COMPLETED。",
            "maxLength": 100,
            "minLength": 1,
            "title": "Workflow Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "goal",
          "workflow_version",
          "status",
          "current_node",
          "created_at",
          "updated_at"
        ],
        "title": "TaskRun",
        "type": "object"
      },
      "TaskRunAction": {
        "enum": [
          "approve",
          "pause",
          "resume",
          "cancel",
          "retry"
        ],
        "title": "TaskRunAction",
        "type": "string"
      },
      "TaskRunActionInput": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TaskRunAction"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "client_token": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Token"
          },
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "option_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Option Id"
          }
        },
        "required": [
          "action"
        ],
        "title": "TaskRunActionInput",
        "type": "object"
      },
      "TaskRunList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskRun"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "TaskRunList",
        "type": "object"
      },
      "Timestamp": {
        "description": "UTC ISO-8601，统一以 Z 结尾。",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,6})?Z$",
        "title": "Timestamp",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "omm_contracts__v1__task_run__Status": {
        "enum": [
          "QUEUED",
          "RUNNING",
          "WAITING_APPROVAL",
          "PAUSED",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "Status",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "OpenMathModel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/task-runs": {
      "get": {
        "operationId": "list_task_runs_api_v1_task_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Task Runs",
        "tags": [
          "task-runs"
        ]
      },
      "post": {
        "operationId": "create_task_run_api_v1_task_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRunInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task Run",
        "tags": [
          "task-runs"
        ]
      }
    },
    "/api/v1/task-runs/{run_id}": {
      "get": {
        "operationId": "get_task_run_api_v1_task_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task Run",
        "tags": [
          "task-runs"
        ]
      }
    },
    "/api/v1/task-runs/{run_id}/actions": {
      "post": {
        "operationId": "post_task_run_action_api_v1_task_runs__run_id__actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRunActionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Task Run Action",
        "tags": [
          "task-runs"
        ]
      }
    },
    "/api/v1/task-runs/{run_id}/approvals": {
      "get": {
        "operationId": "list_task_run_approvals_api_v1_task_runs__run_id__approvals_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Task Run Approvals",
        "tags": [
          "task-runs"
        ]
      }
    },
...
    "/api/v1/task-runs/{run_id}/steps": {
      "get": {
        "operationId": "list_task_run_steps_api_v1_task_runs__run_id__steps_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepRunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Task Run Steps",
        "tags": [
          "task-runs"
        ]
      }
    },
    "/api/v1/task-runs/{run_id}/workspace": {
      "get": {
        "operationId": "get_modeling_workspace_api_v1_task_runs__run_id__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelingWorkspaceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Modeling Workspace",
        "tags": [
          "workspace"
        ]
      }
    },
...
  }
}